Refactored
CSS changes, moved components into their own component folder, added
ngrxstore and updated the clock to work off a service that returns an
observable on a 1 second interaval.
