+++ package.json
@@ -21,6 +21,7 @@
     "@angular/platform-browser": "^4.2.4",
     "@angular/platform-browser-dynamic": "^4.2.4",
     "@angular/router": "^4.2.4",
+    "@ngrx/store": "^4.0.3",
     "angular-resizable-element": "^1.2.3",
     "angular2-draggable": "^1.0.7",
     "core-js": "^2.4.1",
