+++ src/app/core/components/desktop/desktop.component.ts
