+++ src/app/core/components/desktop/icons/icons.component.css
