+++ src/app/core/components/desktop/icons/icons.component.html
