+++ src/app/core/components/desktop/icons/icons.component.ts
