+++ src/app/core/components/taskbar/cortana/cortana.component.css
@@ -24,7 +24,7 @@
 }
 
 .text {
-   color: grey;
+   color: #ADADAD;
    display: inline-block;
    text-align: center;
    font-size: 14px;
