+++ src/app/core/components/taskbar/cortana/cortana.component.ts
