+++ src/app/core/components/taskbar/peek/peek.component.css
