+++ src/app/core/components/taskbar/peek/peek.component.spec.ts
