+++ src/app/core/components/taskbar/start-menu/start-menu.component.css
