+++ src/app/core/components/taskbar/taskview/taskview.component.css
@@ -1,5 +1,5 @@
 .taskview {
-   background: url("../../../../assets/core/task view.png");
+   background: url("../../../../../assets/core/task view.png");
    width: 49px;
    height: 44px;
    background-size: 23px;
@@ -8,7 +8,7 @@
 }
 
 .taskview:hover {
-   background: url("../../../../assets/core/taskviewHover.png");
+   background: url("../../../../../assets/core/taskviewHover.png");
    width: 49px;
    height: 44px;
    background-size: 23px;
