+++ src/app/core/components/taskbar/taskview/taskview.component.ts
