+++ src/app/core/components/taskbar/time/time.component.html
@@ -1,0 +1,4 @@
+<div class="time">
+   <div class="clock">{{ currentTime | date:'shortTime'}}</div>
+   <div class="date">{{ currentTime | date:'shortDate'}}</div>
+</div>
