+++ src/app/core/components/taskbar/tray/tray.component.css
@@ -15,16 +15,16 @@
    transform: rotate(90deg);
    transform-origin: top center;
    position: absolute;
-   left: 20px;
+   left: 25px;
    top: 25px;
    cursor: pointer;
 }
 
 .all-icons {
+   visibility: hidden;
    position:absolute;
    bottom:44px;
    width:150px;
-   margin-left:-100px;
    padding:5px;
    background:rgba(36, 38, 39,1);
    text-align:left;
