+++ src/app/core/components/taskbar/tray/tray.component.html
@@ -1,0 +1,6 @@
+<div class="up-arrow" (click)="trayClick()">
+   <div class="arrow">❮</div>
+</div>
+<div>
+   <div [ngStyle]="getStyle()" class="all-icons">{{getStyle()}}</div>
+</div>
