+++ src/app/core/components/taskbar/tray/tray.component.spec.ts
