+++ src/app/core/components/window/window.component.css
