+++ src/app/core/components/window/window.component.ts
@@ -12,7 +12,6 @@ export interface ResizeObject {
     selector: 'core-window',
     templateUrl: './window.component.html',
     styleUrls: ['./window.component.css'],
-
 })
 export class WindowComponent implements OnInit {
     @Input('settings') settings;
