+++ src/app/core/modules/desktop.module.ts
@@ -5,8 +5,8 @@ import { HttpModule } from '@angular/http';
 import { CommonModule } from '@angular/common';
 import { AngularDraggableModule } from 'angular2-draggable';
 
-import { DesktopComponent } from './desktop.component';
-import { IconsComponent } from './icons/icons.component';
+import { DesktopComponent } from '../components/desktop/desktop.component';
+import { IconsComponent } from '../components/desktop/icons/icons.component';
 
 @NgModule({
     declarations: [
