--- src/app/core/taskbar/notification-menu/notification-menu.component.spec.ts
@@ -1,28 +1,0 @@
-/* tslint:disable:no-unused-variable */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-
-import { NotificationMenuComponent } from './notification-menu.component';
-
-describe('NotificationMenuComponent', () => {
-  let component: NotificationMenuComponent;
-  let fixture: ComponentFixture<NotificationMenuComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ NotificationMenuComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(NotificationMenuComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
