--- src/app/core/taskbar/time/time.component.html
@@ -1,4 +1,0 @@
-<div class="time">
-    <div class="clock">{{ time() | date:'shortTime'}}</div>
-    <div class="date">{{ time() | date:'shortDate'}}</div>
-</div>
