--- src/app/core/taskbar/tray/tray.component.html
@@ -1,6 +1,0 @@
-<div class="up-arrow">
-   <div class="arrow">❮</div>
-</div>
-<div class="">
-   <div class="all-icons">TESTTTTTTTTTTTTT</div>
-</div>
