--- src/app/core/taskbar/tray/tray.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-   selector: 'taskbar-tray',
-   templateUrl: './tray.component.html',
-   styleUrls: ['./tray.component.css']
-})
-export class TrayComponent implements OnInit {
-
-   constructor() { }
-
-   ngOnInit() {
-   }
-
-}
